fix cross-building
