+++ mdoc/src/main/scala/mdoc/internal/markdown/SectionInput.scala
